Reference

Warning

For many layers and functions, the documentation is still missing because this documentation is still under construction!

Reference

GradValley

GradValley.Layers

Containers

Convolution

Pooling

Fully connected

Normalization

Reshape / Flatten

Special activation functions

GradValley.Optimization

GradValley.Functional

GradValley

GradValley.Layers

Containers

Main.GradValley.Layers.SequentialContainer — Type

SequentialContainer(layer_stack::Vector{<: Any})

A sequential container (recommended method for building models). A SequtialContainer can take a vector of layers or other SequentialContainers (submodules). While forward-pass, the given inputs are sequentially propagated through every layer (or submodule) and the output will be returned. The execution order during forward pass is of course the same as the order in the vector containing the layers or submodules. This container currently (!) only accepts Float64 array inputs.

Arguments

layer_stack::Vector{<: Any}: the vector containing the layers (or submodules, so other Containers), the order of the modules in the vector corresponds to the execution order

Examples

# a simple chain of fully connected layers
julia> m = SequentialContainer([Fc(1000, 500), Fc(500, 250), Fc(250, 125)])
# a more complicated example with with nested submodules
julia> feature_extractor_part_1 = SequentialContainer([Conv(1, 6, (5, 5), activation_function="relu"), AvgPool((2, 2))])
julia> feature_extractor_part_2 = SequentialContainer([Conv(6, 16, (5, 5), activation_function="relu"), AvgPool((2, 2))])
julia> feature_extractor = SequentialContainer([feature_extractor_part_1, feature_extractor_part_2])
julia> classifier = SequentialContainer([Fc(256, 120, activation_function="relu"), Fc(120, 84, activation_function="relu"), Fc(84, 10)])
julia> m = SequentialContainer([feature_extractor, Reshape((256, )), classifier, Softmax(dim=2)])
# if a SequentialContainer contains BatchNorm layers (regardless of whether they are nested somewhere in a submodule or not), 
# the mode of all these layers at once can be switched as follows
julia> trainmode!(m)
julia> testmode!(m)
# computing the output of the module (with random inputs)
julia> input = rand(32, 1, 28, 28)
julia> output = forward(m, input)

Convolution

Main.GradValley.Layers.Conv — Type

Conv(in_channels::Int, out_channels::Int, kernel_size::Tuple{Int, Int}; stride::Tuple{Int, Int}=(1, 1), padding::Tuple{Int, Int}=(0, 0), dilation::Tuple{Int, Int}=(1, 1), groups::Int=1, activation_function::Union{Nothing, String}=nothing, init_mode::String="default_uniform", use_bias::Bool=true)

A convolution layer. Apply a 2D convolution over an input signal with additional batch and channel dimensions. This layer currently (!) only accepts Float64 array inputs.

Arguments

in_channels::Int: the number of channels in the input image

out_channels::Int: the number of channels produced by the convolution

kernel_size::Tuple{Int, Int}: the size of the convolving kernel

stride::Tuple{Int, Int}=(1, 1): the stride of the convolution

padding::Tuple{Int, Int}=(0, 0): the zero padding added to all four sides of the input

dilation::Tuple{Int, Int}=(1, 1): the spacing between kernel elements

groups::Int=1: the number of blocked connections from input channels to output channels (in-channels and out-channels must both be divisible by groups)

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the convolution

init_mode::String="default_uniform": the initialization mode of the weights (can be "default_uniform", "default", "kaiming_uniform", "kaiming", "xavier_uniform" or "xavier")

use_bias::Bool=true: if true, adds a learnable bias to the output

Shapes

Input: $(N, C_{in}, H_{in}, W_{in})$

Weight: $(C_{out}, \frac{C_{in}}{groups}, H_{w}, W_{w})$

Bias: $(C_{out}, )$

Output: $(N, C_{out}, H_{out}, W_{out})$, where

$H_{out} = {\frac{H_{in} + 2 \cdot padding[1] - dilation[1] \cdot (H_w - 1)}{stride[1]}} + 1$

$W_{out} = {\frac{W_{in} + 2 \cdot padding[2] - dilation[2] \cdot (W_w - 1)}{stride[2]}} + 1$

Useful Fields/Variables

kernels::Array{Float64, 4}: the learnable weights of the layer

bias::Vector{Float64}: the learnable bias of the layer (used when use_bias=true)

gradients::Array{Float64, 4}: the current gradients of the weights/kernels

bias_gradients::Vector{Float64}: the current gradients of the bias

Definition

For one group, a multichannel 2D convolution (disregarding batch dimension and activation function) can be described as:

$o_{c_{out}, y_{out}, x_{out}} = \big(\sum_{c_{in=1}}^{C_{in}}\sum_{y_w=1}^{H_{w}}\sum_{x_w=1}^{W_{w}} i_{c_{in}, y_{in}, x_{in}} \cdot w_{c_{out}, c_{in}, y_w, x_w}\big) + b_{c_{out}}$, where

$y_{in} = y_{out} + (stride[1] - 1) \cdot (y_{out} - 1) + (y_w - 1) \cdot dilation[1]$

$x_{in} = x_{out} + (stride[2] - 1) \cdot (x_{out} - 1) + (x_w - 1) \cdot dilation[2]$

O is the output array, I the input array, W the weight array and B the bias array.

Examples

# square kernels and fully default values of keyword arguments
julia> m = Conv(3, 6, (5, 5))
# non-square kernels and unequal stride and with padding as well as specified weight initialization mode
# (init_mode="kaiming" stands for kaiming weight initialization with normally distributed values)
julia> m = Conv(3, 6, (3, 5), stride=(2, 1), padding=(2, 1))
# non-square kernels and unequal stride and with padding, dilation and 3 groups
# (because groups=in_channels and out_channles is divisible by groups, it is even a depthwise convolution)
julia> m = Conv(3, 6, (3, 5), stride=(2, 1), padding=(4, 2), dilation=(3, 1), groups=3)
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Main.GradValley.Layers.DepthwiseConv — Type

DepthwiseConv(in_channels::Int, out_channels::Int, kernel_size::Tuple{Int, Int}; stride::Tuple{Int, Int}=(1, 1), padding::Tuple{Int, Int}=(0, 0), dilation::Tuple{Int, Int}=(1, 1), activation_function::Union{Nothing, String}=nothing, init_mode::String="default_uniform", use_bias::Bool=true)

A depthwise convolution layer. Apply a 2D depthwise convolution over an input signal with additional batch and channel dimensions. This layer currently (!) only accepts Float64 array inputs.

Arguments

in_channels::Int: the number of channels in the input image

out_channels::Int: the number of channels produced by the convolution

kernel_size::Tuple{Int, Int}: the size of the convolving kernel

stride::Tuple{Int, Int}=(1, 1): the stride of the convolution

padding::Tuple{Int, Int}=(0, 0): the zero padding added to all four sides of the input

dilation::Tuple{Int, Int}=(1, 1): the spacing between kernel elements

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the convolution

init_mode::String="default_uniform": the initialization mode of the weights (can be "default_uniform", "default", "kaiming_uniform", "kaiming", "xavier_uniform" or "xavier")

use_bias::Bool=true: if true, adds a learnable bias to the output

Shapes

Input: $(N, C_{in}, H_{in}, W_{in})$

Weight: $(C_{out}, \frac{C_{in}}{groups}, H_{w}, W_{w})$, where $groups = in\_channels$

Bias: $(C_{out}, )$

Output: $(N, C_{out}, H_{out}, W_{out})$, where

$H_{out} = {\frac{H_{in} + 2 \cdot padding[1] - dilation[1] \cdot (H_w - 1)}{stride[1]}} + 1$

$W_{out} = {\frac{W_{in} + 2 \cdot padding[2] - dilation[2] \cdot (W_w - 1)}{stride[2]}} + 1$

Useful Fields/Variables

kernels::Array{Float64, 4}: the learnable weights of the layer

bias::Vector{Float64}: the learnable bias of the layer (used when use_bias=true)

gradients::Array{Float64, 4}: the current gradients of the weights/kernels

bias_gradients::Vector{Float64}: the current gradients of the bias

Definition

A convolution is called depthwise if $groups=in\_channels$ and $out\_channels=k \cdot in\_channels$, where $k$ is a positive integer. The second condition ensures that the of number out-channels is divisible by the number of groups/in-channels. In the background, the standard convolution operation is also used for this layer. It is just an interface making clear that this layer can only perform a depthwise convolution.

Examples

# square kernels and fully default values of keyword arguments
julia> m = DepthwiseConv(3, 6, (5, 5))
# non-square kernels and unequal stride and with padding as well as specified weight initialization mode
# (init_mode="kaiming" stands for kaiming weight initialization with normally distributed values)
julia> m = DepthwiseConv(3, 6, (3, 5), stride=(2, 1), padding=(2, 1))
# non-square kernels and unequal stride and with padding, dilation and 3 groups
# (because groups=in_channels and out_channles is divisible by groups, it is even a depthwise convolution)
julia> m = DepthwiseConv(3, 6, (3, 5), stride=(2, 1), padding=(4, 2), dilation=(3, 1), groups=3)
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Pooling

Main.GradValley.Layers.MaxPool — Type

MaxPool(kernel_size::Tuple{Int, Int}; stride::Tuple{Int, Int}=kernel_size, padding::Tuple{Int, Int}=(0, 0), dilation::Tuple{Int, Int}=(1, 1), activation_function::Union{Nothing, String}=nothing)

A maximum pooling layer. Apply a 2D maximum pooling over an input signal with additional batch and channel dimensions. This layer currently (!) only accepts Float64 array inputs.

Arguments

kernel_size::Tuple{Int, Int}: the size of the window to take the maximum over

stride::Tuple{Int, Int}=kernel_size: the stride of the window

padding::Tuple{Int, Int}=(0, 0): the zero padding added to all four sides of the input

dilation::Tuple{Int, Int}=(1, 1): the spacing between the window elements

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the pooling

Shapes

Input: $(N, C, H_{in}, W_{in})$

Output: $(N, C, H_{out}, W_{out})$, where

$H_{out} = {\frac{H_{in} + 2 \cdot padding[1] - dilation[1] \cdot (kernel\_size[1] - 1)}{stride[1]}} + 1$

$W_{out} = {\frac{W_{in} + 2 \cdot padding[2] - dilation[2] \cdot (kernel\_size[2] - 1)}{stride[2]}} + 1$

Definition

A multichannel 2D maximum pooling (disregarding batch dimension and activation function) can be described as:

\[\begin{align*} o_{c, y_{out}, x_{out}} = \max _{y_w = 1, ..., kernel\_size[1] \ x_w = 1, ..., kernel\_size[2]} i_{c, y_{in}, x_{in}} \end{align*}\]

Where

$y_{in} = y_{out} + (stride[1] - 1) \cdot (y_{out} - 1) + (y_w - 1) \cdot dilation[1]$

$x_{in} = x_{out} + (stride[2] - 1) \cdot (x_{out} - 1) + (x_w - 1) \cdot dilation[2]$

O is the output array and I the input array.

Examples

# pooling of square window of size=(3, 3) and automatically selected stride
julia> m = MaxPool((3, 3))
# pooling of non-square window with custom stride and padding
julia> m = MaxPool((3, 2), stride=(2, 1), padding=(1, 1))
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Main.GradValley.Layers.AvgPool — Type

AvgPool(kernel_size::Tuple{Int, Int}; stride::Tuple{Int, Int}=kernel_size, padding::Tuple{Int, Int}=(0, 0), dilation::Tuple{Int, Int}=(1, 1), activation_function::Union{Nothing, String}=nothing)

An average pooling layer. Apply a 2D average pooling over an input signal with additional batch and channel dimensions. This layer currently (!) only accepts Float64 array inputs.

Arguments

kernel_size::Tuple{Int, Int}: the size of the window to take the average over

stride::Tuple{Int, Int}=kernel_size: the stride of the window

padding::Tuple{Int, Int}=(0, 0): the zero padding added to all four sides of the input

dilation::Tuple{Int, Int}=(1, 1): the spacing between the window elements

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the pooling

Shapes

Input: $(N, C, H_{in}, W_{in})$

Output: $(N, C, H_{out}, W_{out})$, where

$H_{out} = {\frac{H_{in} + 2 \cdot padding[1] - dilation[1] \cdot (kernel\_size[1] - 1)}{stride[1]}} + 1$

$W_{out} = {\frac{W_{in} + 2 \cdot padding[2] - dilation[2] \cdot (kernel\_size[2] - 1)}{stride[2]}} + 1$

Definition

A multichannel 2D average pooling (disregarding batch dimension and activation function) can be described as:

$o_{c, y_{out}, x_{out}} = \frac{1}{kernel\_size[1] \cdot kernel\_size[2]} \sum_{i=1}^{kernel\_size[1]}\sum_{j=1}^{kernel\_size[2]} i_{c, y_{in}, x_{in}}$, where

$y_{in} = y_{out} + (stride[1] - 1) \cdot (y_{out} - 1) + (y_w - 1) \cdot dilation[1]$

$x_{in} = x_{out} + (stride[2] - 1) \cdot (x_{out} - 1) + (x_w - 1) \cdot dilation[2]$

O is the output array and I the input array.

Examples

# pooling of square window of size=(3, 3) and automatically selected stride
julia> m = AvgPool((3, 3))
# pooling of non-square window with custom stride and padding
julia> m = AvgPool((3, 2), stride=(2, 1), padding=(1, 1))
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Main.GradValley.Layers.AdaptiveMaxPool — Type

AdaptiveMaxPool(output_size::Tuple{Int, Int}; activation_function::Union{Nothing, String}=nothing)

An adaptive maximum pooling layer. Apply a 2D adaptive maximum pooling over an input signal with additional batch and channel dimensions. For any input size, the size of the output is always equal to the specified $output\_size$. This layer currently (!) only accepts Float64 array inputs.

Arguments

output_size::Tuple{Int, Int}: the target output size of the image (can even be larger than the input size) of the form $(H_{out}, W_{out})$

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the pooling

Shapes

Input: $(N, C, H_{in}, W_{in})$

Output: $(N, C, H_{out}, W_{out})$, where $(H_{out}, W_{out}) = output\_size$

Definition

In some cases, the kernel-size and stride could be calculated in a way that the output would have the target size (using a standard maximum pooling with the calculated kernel-size and stride, padding and dilation would not be used in this case). However, this approach would only work if the input size is an integer multiple of the output size (See this question at stack overflow for further information: stackoverflow.com/questions/53841509/how-does-adaptive-pooling-in-pytorch-work). A more generic approach is to calculate the indices of the input with an additional algorithm only for adaptive pooling. With this approach, it is even possible that the output is larger than the input what is really unusual for pooling simply because that is the opposite of what pooling actually should do, namely reducing the size. The function get_in_indices(in_len, out_len) in gv_functional.jl (line 95 - 113) implements such an algorithm (similar to the one at the stack overflow question), so you could check there on how exactly it is defined. Thus, the mathematical definition would be identical to the one at MaxPool with the difference that the indices $y_{in}$ and $x_{in}$ have already been calculated beforehand.

Examples

# target output size of 5x5
julia> m = AdaptiveMaxPool((5, 5))
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Main.GradValley.Layers.AdaptiveAvgPool — Type

AdaptiveAvgPool(output_size::Tuple{Int, Int}; activation_function::Union{Nothing, String}=nothing)

An adaptive average pooling layer. Apply a 2D adaptive average pooling over an input signal with additional batch and channel dimensions. For any input size, the size of the output is always equal to the specified $output\_size$. This layer currently (!) only accepts Float64 array inputs.

Arguments

output_size::Tuple{Int, Int}: the target output size of the image (can even be larger than the input size) of the form $(H_{out}, W_{out})$

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output after the pooling

Shapes

Input: $(N, C, H_{in}, W_{in})$

Output: $(N, C, H_{out}, W_{out})$, where $(H_{out}, W_{out}) = output\_size$

Definition

In some cases, the kernel-size and stride could be calculated in a way that the output would have the target size (using a standard average pooling with the calculated kernel-size and stride, padding and dilation would not be used in this case). However, this approach would only work if the input size is an integer multiple of the output size (See this question at stack overflow for further information: stackoverflow.com/questions/53841509/how-does-adaptive-pooling-in-pytorch-work). A more generic approach is to calculate the indices of the input with an additional algorithm only for adaptive pooling. With this approach, it is even possible that the output is larger than the input what is really unusual for pooling simply because that is the opposite of what pooling actually should do, namely reducing the size. The function get_in_indices(in_len, out_len) in gv_functional.jl (line 95 - 113) implements such an algorithm (similar to the one at the stack overflow question), so you could check there on how exactly it is defined. Thus, the mathematical definition would be identical to the one at AvgPool with the difference that the indices $y_{in}$ and $x_{in}$ have already been calculated beforehand.

Examples

# target output size of 5x5
julia> m = AdaptiveAvgPool((5, 5))
# computing the output of the layer (with random inputs)
julia> input = rand(32, 3, 50, 50)
julia> output = forward(m, input)

Fully connected

Main.GradValley.Layers.Fc — Type

Fc(in_features::Int, out_features::Int; activation_function::Union{Nothing, String}=nothing, init_mode::String="default_uniform", use_bias::Bool=true)

A fully connected layer (sometimes also known as dense or linear). Apply a linear transformation (matrix multiplication) to the input signal with additional batch dimension. This layer currently (!) only accepts Float64 array inputs.

Arguments

in_features::Int: the size of each input sample ("number of input neurons")

out_features::Int: the size of each output sample ("number of output neurons")

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output

init_mode::String="default_uniform": the initialization mode of the weights (can be "default_uniform", "default", "kaiming_uniform", "kaiming", "xavier_uniform" or "xavier")

use_bias::Bool=true: if true, adds a learnable bias to the output

Shapes

Input: $(N, in\_features)$

Weight: $(out\_features, in\_features)$

Bias: $(out\_features, )$

Output: $(N, out\_features)$

Useful Fields/Variables

weights::Array{Float64, 2}: the learnable weights of the layer

bias::Vector{Float64}: the learnable bias of the layer (used when use_bias=true)

gradients::Array{Float64, 2}: the current gradients of the weights

bias_gradients::Vector{Float64}: the current gradients of the bias

Definition

The forward pass of a fully connected layer is given by the matrix multiplication between the weight matrix and the input vector (disregarding batch dimension and activation function):

$O = WI + B$

This operation can also be described by:

$o_{j} = \big(\sum_{k=1}^{in\_features} w_{j,k} \cdot i_{k}\big) + b_{j}$

O is the output vector, I the input vector, W the weight matrix and B the bias vector. Visually interpreted, it means that each input neuron i is weighted with the corresponding weight w connecting the input neuron to the output neuron o where all the incoming signals are summed up.

Examples

# a fully connected layer with 784 input features and 120 output features
julia> m = Fc(784, 120)
# computing the output of the layer (with random inputs)
julia> input = rand(32, 784)
julia> output = forward(m, input)

Normalization

Main.GradValley.Layers.BatchNorm2d — Type

BatchNorm2d(num_features::Int; epsilon::Float64=1e-05, momentum::Float64=0.1, affine::Bool=true, track_running_stats::Bool=true, activation_function::Union{Nothing, String}=nothing)

A batch normalization layer. Apply a batch normalization over a 4D input signal (a mini-batch of 2D inputs with additional channel dimension). This layer currently (!) only accepts Float64 array inputs.

This layer has two modes: training mode and test mode. If track_running_stats::Bool=true, this layer behaves differently in the two modes. During training, this layer always uses the currently calculated batch statistics. If track_running_stats::Bool=true, the running mean and variance are tracked during training and will be used while testing. If track_running_stats::Bool=false, even in test mode, the currently calculated batch statistics are used. The mode can be switched with trainmode! or testmode! respectively. The training mode is active by default.

Arguments

num_features::Int: the number of channels

epsilon::Float64=1e-05: a value added to the denominator for numerical stability

momentum::Float64=0.1: the value used for the running mean and running variance computation

affine::Bool=true: if true, this layer uses learnable affine parameters/weights ($\gamma$ and $\beta$)

track_running_stats::Bool=true: if true, this layer tracks the running mean and variance during training and will use them for testing/evaluation, if false, such statistics are not tracked and, even in test mode, the batch statistics are always recalculated for each new input

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output

Shapes

Input: $(N, C, H, W)$

$\gamma$ Weight, $\beta$ Bias: $(C, )$

Running Mean/Variance: $(C, )$

Output: $(N, C, H, W)$ (same shape as input)

Useful Fields/Variables

Weights (used if affine::Bool=true)

weight_gamma::Vector{Float64}: $\gamma$, a learnabele parameter for each channel, initialized with ones

weight_beta::Vector{Float64}: $\beta$, a learnabele parameter for each channel, initialized with zeros

Gradients of weights (used if affine::Bool=true)

gradient_gamma::Vector{Float64}: the gradients of $\gamma$

gradient_beta::Vector{Float64}: the gradients of $\beta$

Running statistics (used if rack_running_stats::Bool=true)

running_mean::Vector{Float64}: the continuously updated batch statistics of the mean

running_variance::Vector{Float64}: the continuously updated batch statistics of the variance

Definition

A batch normalization operation can be described as: For input values over a mini-batch: $\mathcal{B} = \{x_1, x_2, ..., x_n\}$

\[\begin{align*} y_i = \frac{x_i - \overline{\mathcal{B}}}{\sqrt{Var(\mathcal{B}) + \epsilon}} \cdot \gamma + \beta \end{align*}\]

Where $y_i$ is an output value and $x_i$ an input value. $\overline{\mathcal{B}}$ is the mean of the input values in $\mathcal{B}$ and $Var(\mathcal{B})$ is the variance of the input values in $\mathcal{B}$. Note that this definition is fairly general and not specified to 4D inputs. In this case, the input values of $\mathcal{B}$ are taken for each channel individually. So the mean and variance are calculated per channel over the mini-batch.

The update rule for the running statistics (running mean/variance) is:

\[\begin{align*} \hat{x}_{new} = (1 - momentum) \cdot \hat{x} + momentum \cdot x \end{align*}\]

Where $\hat{x}$ is the estimated statistic and $x$ is the new observed value. So $\hat{x}_{new}$ is the new, updated estimated statistic.

Examples

# a batch normalization layer (3 channels) with learnabel parameters and continuously updated batch statistics for evaluation
julia> m = BatchNorm2d(3)
# the mode can be switched with trainmode! or testmode!
julia> trainmode!(m)
julia> testmode!(m)
# computing the output of the layer (with random inputs)
julia> input = rand(32, 1, 28, 28)
julia> output = forward(m, input)

Reshape / Flatten

Main.GradValley.Layers.Reshape — Type

Reshape(out_shape; activation_function::Union{Nothing, String}=nothing)

A reshape layer (probably mostly used as a flatten layer). Reshape the input signal (effects all dimensions except the batch dimension). This layer currently (!) only accepts Float64 array inputs.

Arguments

out_shape: the target output size (the output has the same data as the input and must have the same number of elements)

activation_function::Union{Nothing, String}=nothing: the element-wise activation function which will be applied to the output

Shapes

Input: $(N, *)$, where * means any number of dimensions

Output: $(N, out\_shape...)$

Definition

This layer uses the standard reshape function inbuilt in Julia.

Examples

# flatten the input of size 1*28*28 to a vector of length 784 (each plus batch dimension of course)
julia> m = Reshape((784, ))
# computing the output of the layer (with random inputs)
julia> input = rand(32, 1, 28, 28)
julia> output = forward(m, input)

Special activation functions

Missing docstring.

Missing docstring for Softmax. Check Documenter's build log for details.

GradValley.Optimization

GradValley.Functional